requirements.txt
Flask>=3.1
Flask-Login>=0.6
Flask-SQLAlchemy>=3.1
APScheduler>=3.11
python-dotenv>=1.0
netmiko>=4.4
pyzabbix>=1.3
requests>=2.31
